### 2021-07-07 #42 @shailesh1729
CR-Sparse: XLA accelerated algorithms for inverse problems in sparse representations and compressive sensing 
Submitting Author: Shailesh Kumar (@shailesh1729)
Package Name: CR-Sparse
One-Line Description of Package:  XLA accelerated algorithms for inverse problems in sparse representations and compressive sensing
Repository Link (if existing):   https://github.com/carnotresearch/cr-sparse
Documentation: https://carnotresearch.github.io/cr-sparse/

---

This library aims to provide XLA/JAX based Python implementations for
various algorithms related to:
* Sparse approximation
* Compressive sensing

The library also provides
* Various simple dictionaries and sensing matrices
* Sample data generation utilities
* Framework for evaluation of sparse recovery algorithms

Addition information
* [An evaluation of the sparse recovery performance of algorithms included in the package](https://nbviewer.jupyter.org/github/carnotresearch/cr-sparse/blob/master/experiments/pursuit/comparison/nb_recovery_comparison_multiple_methods.ipynb)
* [Some speed benchmarks](https://carnotresearch.github.io/cr-sparse/benchmarks/comparison.html)

The sparse recovery algorithms reconstruct sparse representations of a signal or a set of measurements in a dictionary (which is an overcomplete basis). The closest category I can think of is data munging.

Compressive sensing has applications in areas like photography, holography, facial recognition, magnetic resonance imaging, etc.
Sparse representations have applications in areas like image deblurring, denoising, dictionary learning, image compression, inpainting etc.

[scikit-learn](https://scikit-learn.org/stable/modules/linear_model.html) contains linear models which include implementations of algorithms like orthogonal matching pursuit (OMP), LASSO, LARS, OLS, etc.

CR-Sparse implements these algorithms using a functional programming paradigm built on top of [JAX](https://jax.readthedocs.io/). This enables it to use immense optimization opportunities provided by the [XLA](https://www.tensorflow.org/xla) compiler. In addition, the code can be deployed on GPUs and TPUs without any modification. These are significant advantages compared to the scikit-learn implementation.

The coverage of greedy sparse recovery algorithms in this package is bigger. E.g. it includes implementations of hard thresholding pursuit, compressive sampling matching pursuit, iterative hard thresholding, etc. in a single package.

We are also adding implementations for L1 minimization (convex optimization) based algorithms.

This is a pre-submission enquiry to understand if the CR-sparse package described below fits the submission criteria for pyOpenSci. Please do let us know what more should be done.
