### 2021-07-07 #42 @NickleDave
Hi @shailesh1729 thank you for your interest in pyOpenSci.

I actually work with compressive sensing in my day job and think it's great to see some work on this in the Python space.

I do think this is out of scope for pyOpenSci though.
We are mainly focused on tooling for open science.
E.g., if you built a tool that made it easier to download datasets from a repository of MRI images, say, to test compressive sensing algorithms or whatever else you wanted to do with it, that would be in scope (that would be "data retrieval").

Have you considered submission at JOSS?
https://joss.theoj.org/
