### 2021-07-07 #42 @NickleDave
This is off-topic for the question of scope, but some related repos just in case you're not aware of them already:

* https://github.com/nikcleju/pyCSalgos
* https://github.com/SebastianSemper/camiba
* https://github.com/SIP-AAU/Magni
* https://github.com/ngcthuong/Reproducible-Deep-Compressive-Sensing
* https://github.com/dedupeio/pylbfgs <-- nice Python wrapper around LBFGS that let's you solve L1-constrained problems with OWL-QN

I get that you have different goals than most of them--building on top of XLA/JAX, which is great--but you might be curious about previous attempts at APIs, algorithms, etc.
