### 2021-07-07 #42 @palacio-tech
@NickleDave  Thank you for your feedback. Now I better understand the scope of pyOpenSci. I will try to submit it to JOSS.

Thank you for the references for previous works. Yes, I am slowly implementing these ideas in XLA/JAX. I was aware of some of the previous work but not all of them. This will be helpful.
